requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
requests
psycopg2
